+++ examples/Python/geometry/kd_tree_search.py
@@ -1,8 +1,10 @@
-# cloudViewer: Asher-1.github.io
-# The MIT License (MIT)
-# See license file or visit Asher-1.github.io for details
-
-# examples/Python/Basic/kdtree.py
+# ----------------------------------------------------------------------------
+# -                        CloudViewer: www.cloudViewer.org                  -
+# ----------------------------------------------------------------------------
+# Copyright (c) 2018-2024 www.cloudViewer.org
+# SPDX-License-Identifier: MIT
+# ----------------------------------------------------------------------------
+"""Build a KDTree and use it for neighbour search"""
 
 import numpy as np
 import cloudViewer as cv3d