+++ examples/Python/reconstruction_system/debug/pairwise_pc_alignment.py
@@ -1,15 +1,17 @@
-# CloudViewer: asher-1.github.io
-# The MIT License (MIT)
-# See license file or visit Asher-1.github.io for details
-
-# examples/Python/ReconstructionSystem/debug/pairwise_pc_alignment.py
+# ----------------------------------------------------------------------------
+# -                        CloudViewer: www.cloudViewer.org                  -
+# ----------------------------------------------------------------------------
+# Copyright (c) 2018-2024 www.cloudViewer.org
+# SPDX-License-Identifier: MIT
+# ----------------------------------------------------------------------------
 
 import argparse
 import json
 import sys
 
-sys.path.append("../utility")
-from file import *
+pyexample_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+sys.path.append(pyexample_path)
+from cloudViewer_example import *
 
 sys.path.append(".")
 from initialize_config import *