+++ examples/Python/visualization/load_save_viewpoint.py
@@ -1,10 +1,10 @@
-# cloudViewer: Asher-1.github.io
-# The MIT License (MIT)
-# See license file or visit Asher-1.github.io for details
+# ----------------------------------------------------------------------------
+# -                        CloudViewer: www.cloudViewer.org                  -
+# ----------------------------------------------------------------------------
+# Copyright (c) 2018-2024 www.cloudViewer.org
+# SPDX-License-Identifier: MIT
+# ----------------------------------------------------------------------------
 
-# examples/Python/Advanced/load_save_viewpoint.py
-
-import numpy as np
 import cloudViewer as cv3d