+++ libs/cloudViewer/visualization/utility/DrawGeometry.cpp
@@ -40,7 +40,7 @@ bool DrawGeometries(
         Eigen::Vector3d *lookat /* = nullptr */,
         Eigen::Vector3d *up /* = nullptr */,
         Eigen::Vector3d *front /* = nullptr */,
-        double *zoom /* = zoom */) {
+        double *zoom /* = nullptr */) {
     Visualizer visualizer;
     if (!visualizer.CreateVisualizerWindow(window_name, width, height, left,
                                            top)) {