+++ python/test/cloudViewer_test.py
@@ -54,17 +54,3 @@ def list_devices_with_torch():
             return [cv3d.core.Device("CPU:0")]
     else:
         return []
-
-
-def download_fountain_dataset():
-    fountain_path = os.path.join(test_data_dir, "fountain_small")
-    fountain_zip_path = os.path.join(test_data_dir, "fountain.zip")
-    if not os.path.exists(fountain_path):
-        print("Downloading fountain dataset")
-        url = "https://storage.googleapis.com/isl-datasets/open3d-dev/fountain.zip"
-        urllib.request.urlretrieve(url, fountain_zip_path)
-        print("Extracting fountain dataset")
-        with zipfile.ZipFile(fountain_zip_path, "r") as zip_ref:
-            zip_ref.extractall(os.path.dirname(fountain_path))
-        os.remove(fountain_zip_path)
-    return fountain_path